+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 <a name="unreleased"></a>
 ## [Unreleased]
 
+### Added
+- `advanced_plus.php` example
+
+### Renamed
+- `SpinnerSumbols` to `Symbols`
+
 
 <a name="0.15.1-ALPHA.2"></a>
 ## [0.15.1-ALPHA.2] - 2019-08-13